+++ R/mirai-package.R
@@ -19,13 +19,24 @@
 #'     (\href{https://orcid.org/0000-0002-0750-061X}{ORCID})
 #'
 #' @importFrom nanonext call_aio context is_nul_byte recv_ctx request send_aio
-#'     send_ctx socket stop_aio
+#'     send_ctx socket stop_aio .mirai_scm
 #' @importFrom stats runif
 #'
 #' @docType package
 #' @name mirai-package
 #'
 NULL
+
+.onLoad <- function(libname, pkgname) {
+  mirai <- mirai()
+  mirai <<- mirai
+  invisible()
+}
+
+.onUnload <- function(libpath) {
+  mirai(0L)
+  invisible()
+}
 
 #' @export
 nanonext::is_nul_byte