+++ tests/tests.R
@@ -6,7 +6,7 @@ mirai <- eval_mirai({
 }, m = 2L, n = 3L)
 invisible(!is.null(mirai$aio) || stop())
 invisible(!is.null(mirai$socket) || stop())
-invisible(!is.null(call_mirai(mirai)) || stop())
+invisible(inherits(call_mirai(mirai), "mirai") || stop())
 invisible(mirai$value == 3L || stop())
 invisible(is.null(mirai$aio) || stop())
 invisible(is.null(mirai$socket) || stop())