+++ NEWS.md
@@ -1,7 +1,8 @@
-# mirai 0.1.0.9001 (development)
+# mirai 0.1.0.9002 (development)
 
 * `call_mirai()` gains an argument 'wait'. Choose whether to wait for an async operation to complete or else return immediately.
 * `eval_mirai()` adds support for evaluating arbitrary length expressions wrapped in `{}`.
+* New `mirai()` daemon manager for persistent background processes [experimental].
 
 # mirai 0.1.0